fix : 챌린지 상세조회 달성률 double로 변경

+++ backend/challenge-service/src/main/java/com/ssafy/challengeservice/dto/ChallengeDetailRes.java
@@ -41,7 +41,7 @@ public class ChallengeDetailRes {
         challengeDetailRes.participantsCnt = challenge.getParticipantsCnt();
         challengeDetailRes.rewardPoint = challenge.getRewardPoint();
         challengeDetailRes.progress = challenge.getProgress();
-        challengeDetailRes.progressRate = (challenge.getProgress() / challenge.getGoal()) * 100;
+        challengeDetailRes.progressRate = ((double)challenge.getProgress() / challenge.getGoal()) * 100;
         challengeDetailRes.finishFlag = challenge.getFinishFlag();
         challengeDetailRes.imageUrl = challenge.getImageUrl();
         challengeDetailRes.memberId = challenge.getMember().getId().toString();
